% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{add_link}
\alias{add_link}
\title{Add a link to a STAC catalog}
\usage{
add_link(catalog, rel, href, ...)
}
\arguments{
\item{catalog}{A STAC catalog, collection, or item object.}

\item{rel}{(character, required) The link relation type. Common values include
\code{"self"}, \code{"root"}, \code{"parent"}, \code{"child"}, and \code{"item"}. See the STAC
specification for a full list of relation types.}

\item{href}{(character, required) The URL or path to the linked resource.
Can be absolute or relative.}

\item{...}{Additional link properties passed to \code{stac_link()}, such as \code{type},
\code{title}, \code{method}, \code{headers}, \code{body}, or \code{merge}.}
}
\value{
The modified catalog object with the new link added.
}
\description{
Adds a link object to a STAC Catalog, Collection, or Item. Links are used
to connect STAC resources and provide relationships between catalogs, collections,
and items.
}
\examples{
catalog <- stac_catalog(
  id = "my-catalog",
  description = "Example catalog"
)

# Add a self link
catalog <- add_link(
  catalog,
  rel = "self",
  href = "https://example.com/catalog.json",
  type = "application/json"
)

# Add a related link with a title
catalog <- add_link(
  catalog,
  rel = "related",
  href = "https://example.com/metadata.html",
  type = "text/html",
  title = "Additional metadata"
)

}
\seealso{
\itemize{
\item \code{\link[=add_self_link]{add_self_link()}} for adding a self link
\item \code{\link[=add_root_link]{add_root_link()}} for adding a root link
\item \code{\link[=add_parent_link]{add_parent_link()}} for adding a parent link
\item \code{\link[=add_child]{add_child()}} for adding a child catalog or collection
}
}
