% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{add_parent_link}
\alias{add_parent_link}
\title{Add a parent link to a STAC catalog}
\usage{
add_parent_link(catalog, href)
}
\arguments{
\item{catalog}{A STAC catalog, collection, or item object.}

\item{href}{(character, required) The URL to the parent catalog or collection.
Can be absolute or relative.}
}
\value{
The modified catalog object with the parent link added.
}
\description{
Adds a parent link to a STAC Catalog, Collection, or Item. A parent link provides
the URL to the parent catalog or collection in the STAC hierarchy. Non-root
catalogs should include a parent link.
}
\examples{
catalog <- stac_catalog(
  id = "child-catalog",
  description = "A child catalog"
)

catalog <- add_parent_link(catalog, "../parent/catalog.json")

}
\seealso{
\itemize{
\item \code{\link[=add_self_link]{add_self_link()}} for adding a self link
\item \code{\link[=add_root_link]{add_root_link()}} for adding a root link
\item \code{\link[=add_link]{add_link()}} for adding arbitrary links
}
}
