% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{add_projection_metadata}
\alias{add_projection_metadata}
\title{Add Projection Extension Metadata}
\usage{
add_projection_metadata(item, r, asset_key)
}
\arguments{
\item{item}{A STAC Item object.}

\item{r}{A SpatRaster object.}

\item{asset_key}{The asset key to add projection metadata to.}
}
\value{
The modified STAC Item.
}
\description{
Adds projection extension metadata to a STAC Item for rasters not in WGS84.
}
\keyword{internal}
