% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{bbox_from_sf}
\alias{bbox_from_sf}
\title{Calculate Bounding Box from sf Object}
\usage{
bbox_from_sf(sf_obj)
}
\arguments{
\item{sf_obj}{An sf object.}
}
\value{
A numeric vector of length 4: c(west, south, east, north).
}
\description{
Calculates a bounding box from an sf object in the format required by STAC.
}
