% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{extent_from_items}
\alias{extent_from_items}
\title{Create Collection Extent from Multiple Items}
\usage{
extent_from_items(items)
}
\arguments{
\item{items}{A list of STAC Item objects.}
}
\value{
A list with spatial and temporal extent suitable for \code{stac_collection()}.
}
\description{
Calculates the spatial and temporal extent for a collection from a list of items.
}
\examples{
\dontrun{
items <- list(item1, item2, item3)
extent <- extent_from_items(items)

collection <- stac_collection(
  id = "my-collection",
  description = "Collection of items",
  license = "CC0-1.0",
  extent = extent
)
}

}
