% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item-helpers.R
\name{find_link}
\alias{find_link}
\title{Find a Link by Relation Type}
\usage{
find_link(stac_object, rel)
}
\arguments{
\item{stac_object}{A STAC Catalog, Collection, or Item object.}

\item{rel}{The link relation type to find (e.g., "self", "root", "parent").}
}
\value{
The first link object with the matching rel type, or NULL if not found.
}
\description{
Helper function to find a link with a specific relation type in a STAC object.
}
\keyword{internal}
