% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrations.R
\name{get_media_type}
\alias{get_media_type}
\title{Get Media Type for File}
\usage{
get_media_type(file)
}
\arguments{
\item{file}{File path.}
}
\value{
Media type string.
}
\description{
Determines the appropriate MIME type for a file based on extension.
}
\keyword{internal}
