% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-eo.R
\name{landsat_oli_bands}
\alias{landsat_oli_bands}
\title{Create Standard Landsat 8/9 OLI Bands}
\usage{
landsat_oli_bands(include_thermal = FALSE)
}
\arguments{
\item{include_thermal}{(logical, optional) If TRUE, includes TIRS thermal
bands (B10, B11). Default is FALSE (OLI bands only).}
}
\value{
A list of EO band objects representing Landsat OLI/TIRS bands.
}
\description{
Helper function to create standard band definitions for Landsat 8 and 9
Operational Land Imager (OLI) sensors.
}
\examples{
bands <- landsat_oli_bands()

item <- item |>
  add_eo_extension(bands = bands)

}
