% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-eo.R
\name{print.eo_band}
\alias{print.eo_band}
\title{Print method for EO band objects}
\usage{
\method{print}{eo_band}(x, ...)
}
\arguments{
\item{x}{An EO band object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for EO band objects
}
