% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-raster.R
\name{raster_statistics}
\alias{raster_statistics}
\title{Create Raster Statistics Object}
\usage{
raster_statistics(
  minimum = NULL,
  maximum = NULL,
  mean = NULL,
  stddev = NULL,
  valid_percent = NULL
)
}
\arguments{
\item{minimum}{(numeric, optional) Minimum pixel value in the band.}

\item{maximum}{(numeric, optional) Maximum pixel value in the band.}

\item{mean}{(numeric, optional) Mean (average) pixel value in the band.}

\item{stddev}{(numeric, optional) Standard deviation of pixel values.}

\item{valid_percent}{(numeric, optional) Percentage of valid (non-nodata)
pixels. Should be between 0 and 100.}
}
\value{
A list representing a statistics object.
}
\description{
Creates a statistics object for describing the distribution of pixel values
in a raster band.
}
\examples{
stats <- raster_statistics(
  minimum = 0,
  maximum = 10000,
  mean = 2500,
  stddev = 1200,
  valid_percent = 99.8
)

}
