% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extension-eo.R
\name{sentinel2_msi_bands}
\alias{sentinel2_msi_bands}
\title{Create Standard Sentinel-2 MSI Bands}
\usage{
sentinel2_msi_bands()
}
\value{
A list of EO band objects representing Sentinel-2 MSI bands.
}
\description{
Helper function to create standard band definitions for Sentinel-2
MultiSpectral Instrument (MSI) sensors.
}
\examples{
bands <- sentinel2_msi_bands()

item <- item |>
  add_eo_extension(bands = bands)

}
